Follow-up layer 224 fixes: peer_id as long long, channel support, query fixes
tgl.h: Change tgl_get_peer_id/tgl_set_peer_id to use long long instead
of int for peer IDs (required for channel IDs which exceed int range).
Fix tgl_cmp_peer_id to do proper field-by-field comparison instead of
memcmp (endianness-safe).

binlog.c: Fix bl_do_user to handle long long user IDs properly.
Remove asserts that fail when peers are already created.

queries.c: Clean up debug logging, fix flag bit for channel messages
in messages.sendMessage (bit 4 instead of bit 13), fix various
query parameter ordering issues.

structures.c: Minor fixes for new type structures.

tgl-binlog.h, tgl-layout.h: Updated for new API types.

updates.c: Fix channel ID formatting in log messages (%lld instead of %d).

+++ structures.c
@@ -1642,7 +1642,7 @@ void tglp_insert_channel (struct tgl_state *TLS, tgl_peer_t *P) {
   TLS->Peers[TLS->peer_num ++] = P;
 }
 
-void tgl_insert_empty_user (struct tgl_state *TLS, int uid) {
+void tgl_insert_empty_user (struct tgl_state *TLS, long long uid) {
   tgl_peer_id_t id = TGL_MK_USER (uid);
   if (tgl_peer_get (TLS, id)) { return; }
   tgl_peer_t *P = talloc0 (sizeof (*P));
@@ -1650,7 +1650,7 @@ void tgl_insert_empty_user (struct tgl_state *TLS, int uid) {
   tglp_insert_user (TLS, P);
 }
 
-void tgl_insert_empty_chat (struct tgl_state *TLS, int cid) {
+void tgl_insert_empty_chat (struct tgl_state *TLS, long long cid) {
   tgl_peer_id_t id = TGL_MK_CHAT (cid);
   if (tgl_peer_get (TLS, id)) { return; }
   tgl_peer_t *P = talloc0 (sizeof (*P));
